refactor(spec,mobile): reduce package complexity
